+++ paths/path.go
@@ -3,8 +3,6 @@ package paths
 import (
 	"os"
 	"path/filepath"
-	"reflect"
-	"syscall"
 	"time"
 
 	"github.com/stkali/utility/errors"
@@ -13,7 +11,7 @@ import (
 
 var (
 	ToAbsPath = tool.ToAbsPath
-	UserHome = tool.UserHome
+	UserHome  = tool.UserHome
 )
 
 // GetFileCreated get the creation time of the file through the file name.
@@ -22,28 +20,28 @@ func GetFileCreated(file string) (t time.Time, err error) {
 	if err != nil {
 		return t, errors.Newf("failed to open file: %s, err: %s", file, err)
 	}
-	return GetFdCreated(info)
+	return GetFdCreated(info), nil
 }
 
-// GetFdCreated get the creation time of the file through the fd *os.FileInfo.
-func GetFdCreated(fd os.FileInfo) (time.Time, error) {
-	st := fd.Sys().(*syscall.Stat_t)
-	stValue := reflect.ValueOf(st).Elem()
+// // GetFdCreated get the creation time of the file through the fd *os.FileInfo.
+// func GetFdCreated(fd os.FileInfo) (time.Time, error) {
+// 	st := fd.Sys().(*syscall.Stat_t)
+// 	stValue := reflect.ValueOf(st).Elem()
 
-	// linux
-	cTimeValue := stValue.FieldByName("Ctim")
-	if cTimeValue.Kind() != reflect.Invalid {
-		timeSpec := cTimeValue.Interface().(syscall.Timespec)
-		return time.Unix(timeSpec.Sec, timeSpec.Nsec), nil
-	}
-	// mac
-	cTimeValue = stValue.FieldByName("Ctimespec")
-	if cTimeValue.Kind() != reflect.Invalid {
-		timeSpec := cTimeValue.Interface().(syscall.Timespec)
-		return time.Unix(timeSpec.Sec, timeSpec.Nsec), nil
-	}
-	return time.Time{}, errors.Newf("failed to get file created time")
-}
+// 	// linux
+// 	cTimeValue := stValue.FieldByName("Ctim")
+// 	if cTimeValue.Kind() != reflect.Invalid {
+// 		timeSpec := cTimeValue.Interface().(syscall.Timespec)
+// 		return time.Unix(timeSpec.Sec, timeSpec.Nsec), nil
+// 	}
+// 	// mac
+// 	cTimeValue = stValue.FieldByName("Ctimespec")
+// 	if cTimeValue.Kind() != reflect.Invalid {
+// 		timeSpec := cTimeValue.Interface().(syscall.Timespec)
+// 		return time.Unix(timeSpec.Sec, timeSpec.Nsec), nil
+// 	}
+// 	return time.Time{}, errors.Newf("failed to get file created time")
+// }
 
 func SplitWithExt(path string) (string, string, string) {
 	vol := filepath.VolumeName(path)